This is my new update
It prints the full date (Sunday November 30) for each day. it does not change the date counter properly not stop printing properly yet, but it's pretty solid. Also, this now auto changes the colors for the month variable that is passed in. I have initialized a bunch of variables such as month startD startY and such in the methods but they will ultimately be passed in prom a scanner user input in the final product